feat: push resizeContainer action on it's own file
