+++ src/lib/Branch/index.svelte
@@ -7,6 +7,7 @@
 	import Group from '$lib/primitives/Group.svelte';
 	import { flip } from 'svelte/animate';
 	import { fly } from 'svelte/transition';
+	import { resizeContainer } from '$lib/actions/resizeContainer';
 
 	export let data: IBranch;
 	export let selected = false;
@@ -75,33 +76,6 @@
 		'default',
 		'trunk'
 	];
-
-	// https://svelte.dev/repl/5b46f83924f5437dae097b612f3d97b0?version=3.38.2
-	function resizeContainer(node: HTMLElement) {
-		function updateHeight() {
-			const childrenHeight = (Array.from(node.children) as HTMLElement[]).reduce(
-				(acc, item) => acc + item.offsetHeight,
-				0
-			);
-			node.style.setProperty('height', `${childrenHeight}px`);
-		}
-
-		const observer = new MutationObserver((mutations) => {
-			updateHeight();
-		});
-		observer.observe(node, { characterData: true, subtree: true, childList: true });
-
-		updateHeight();
-
-		node.style.setProperty('overflow', 'hidden');
-		node.style.setProperty('transition', 'height 150ms ease');
-
-		return {
-			destroy() {
-				observer?.disconnect();
-			}
-		};
-	}
 </script>
 
 <div
